+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Improved performance serializing profiling data (#2863)
 - Possible crash in Core Data tracking (#2865)
 - Ensure the current GPU frame rate is always reported for concurrent transaction profiling metrics (#2929)
+- Move profiler metric collection to a background queue (#2956)
 
 ## 8.5.0